+++ spec/mgmt_console_spec.rb
@@ -10,5 +10,4 @@ describe 'rabbitmq::mgmt_console' do
   it 'includes the `default` recipe' do
     expect(chef_run).to include_recipe('rabbitmq::default')
   end
-
 end