+++ test/integration/lwrps/rspec/lwrps_spec.rb
@@ -1,10 +1,9 @@
 require 'spec_helper'
 
-describe file('/usr/local/bin/rabbitmqadmin') do
-  it { should be_file }
-  it { should be_executable }
+describe command('rabbitmqadmin --version') do
+  it { should return_exit_status 0 }
 end
 
-describe command('/usr/local/bin/rabbitmqadmin --version') do
-  it { should return_exit_status 0 }
+describe command('rabbitmqctl list_policies') do
+  its(:stdout) { should match /\/\s+rabbitmq_cluster\s+queues\s+cluster\.\*\s+{"ha-mode":"all","ha-sync-mode":"automatic"}\s+0/ } # rubocop:disable all
 end